move xml_boilerplate.py and refactor tests
